Factorize test code from test_push_github and test_push_gitlab
