+++ tsrc/test/cli/test_push_github.py
@@ -7,26 +7,7 @@ import pytest
 import tsrc.git
 from tsrc.cli.push import RepositoryInfo
 from tsrc.cli.push_github import PushAction
-
-
-@pytest.fixture
-def foo_path(monkeypatch, git_server, tsrc_cli, workspace_path):
-    """ Path to a freshly cloned repository """
-    git_server.add_repo("foo/bar")
-    manifest_url = git_server.manifest_url
-    tsrc_cli.run("init", manifest_url)
-    foo_path = workspace_path.joinpath("foo/bar")
-    monkeypatch.chdir(foo_path)
-
-
-@pytest.fixture
-def push_args():
-    args = types.SimpleNamespace()
-    args.force = False
-    args.title = None
-    args.target_branch = "master"
-    args.merge = False
-    return args
+from tsrc.test.helpers.push import repo_path, push_args
 
 
 @pytest.fixture
@@ -35,14 +16,14 @@ def github_mock():
     return github_mock
 
 
-def execute_push(foo_path, push_args, github_api):
+def execute_push(repo_path, push_args, github_api):
     repository_info = RepositoryInfo()
-    repository_info.read_working_path(foo_path)
+    repository_info.read_working_path(repo_path)
     push_action = PushAction(repository_info, push_args, github_api=github_api)
     push_action.execute()
 
 
-def test_create(foo_path, tsrc_cli, github_mock, push_args):
+def test_create(repo_path, tsrc_cli, github_mock, push_args):
     stub_repo = mock.Mock()
     stub_repo.iter_pulls.return_value = list()
     github_mock.repository.return_value = stub_repo
@@ -50,17 +31,17 @@ def test_create(foo_path, tsrc_cli, github_mock, push_args):
     stub_repo.create_pull.return_value = stub_pr
     stub_pr.html_url = "https://github.com/foo/bar/pull/42"
 
-    tsrc.git.run_git(foo_path, "checkout", "-b", "new-feature")
-    tsrc.git.run_git(foo_path, "commit", "--message", "new feature", "--allow-empty")
+    tsrc.git.run_git(repo_path, "checkout", "-b", "new-feature")
+    tsrc.git.run_git(repo_path, "commit", "--message", "new feature", "--allow-empty")
     push_args.title = "new feature"
-    execute_push(foo_path, push_args, github_mock)
+    execute_push(repo_path, push_args, github_mock)
 
     github_mock.repository.assert_called_with("foo", "bar")
     stub_repo.iter_pulls.assert_called_with()
     stub_repo.create_pull.assert_called_with("new feature", "master", "new-feature")
 
 
-def test_merge(foo_path, tsrc_cli, github_mock, push_args):
+def test_merge(repo_path, tsrc_cli, github_mock, push_args):
     closed_pr = mock.Mock()
     closed_pr.number = 1
     closed_pr.state = "closed"
@@ -81,10 +62,10 @@ def test_merge(foo_path, tsrc_cli, github_mock, push_args):
     stub_repo.iter_pulls.return_value = [closed_pr, opened_pr, opened_pr_wrong_branch]
     github_mock.repository.return_value = stub_repo
 
-    tsrc.git.run_git(foo_path, "checkout", "-b", "new-feature")
-    tsrc.git.run_git(foo_path, "commit", "--message", "new feature", "--allow-empty")
+    tsrc.git.run_git(repo_path, "checkout", "-b", "new-feature")
+    tsrc.git.run_git(repo_path, "commit", "--message", "new feature", "--allow-empty")
     push_args.merge = True
-    execute_push(foo_path, push_args, github_mock)
+    execute_push(repo_path, push_args, github_mock)
 
     stub_repo.create_pull.assert_not_called()
     opened_pr.merge.assert_called_with()
