+++ tsrc/test/cli/test_push_gitlab.py
@@ -7,6 +7,7 @@ import tsrc.git
 import tsrc.gitlab
 from tsrc.cli.push import RepositoryInfo
 from tsrc.cli.push_gitlab import PushAction
+from tsrc.test.helpers.push import repo_path, push_args
 
 GITLAB_URL = "http://gitlab.example.com"
 TIMOTHEE = {"name": "timothee", "id": 1}
@@ -19,30 +20,6 @@ MR_STUB = {"id": "3978", "iid": "42", "web_url": "http://mr/42", "title": "Borin
 PROJECT_IDS = {
     "foo/bar": "42"
 }
-
-
-@pytest.fixture
-def foo_path(monkeypatch, git_server, tsrc_cli, workspace_path):
-    """ Path to a freshly cloned repository """
-    git_server.manifest.configure_gitlab(url=GITLAB_URL)
-    git_server.add_repo("foo/bar")
-    manifest_url = git_server.manifest_url
-    tsrc_cli.run("init", manifest_url)
-    foo_path = workspace_path.joinpath("foo/bar")
-    monkeypatch.chdir(foo_path)
-
-
-@pytest.fixture
-def push_args():
-    args = types.SimpleNamespace()
-    args.accept = False
-    args.target_branch = "master"
-    args.mr_title = None
-    args.assignee = None
-    args.force = False
-    args.ready = None
-    args.wip = None
-    return args
 
 
 @pytest.fixture
@@ -72,16 +49,16 @@ def gitlab_mock():
     return gl_mock
 
 
-def execute_push(foo_path, push_args, gitlab_mock):
+def execute_push(repo_path, push_args, gitlab_mock):
     repository_info = RepositoryInfo()
-    repository_info.read_working_path(foo_path)
+    repository_info.read_working_path(repo_path)
     push_action = PushAction(repository_info, push_args, gl_helper=gitlab_mock)
     push_action.execute()
 
 
-def test_creating_merge_request(foo_path, tsrc_cli, gitlab_mock, push_args):
-    tsrc.git.run_git(foo_path, "checkout", "-b", "new-feature")
-    tsrc.git.run_git(foo_path, "commit", "--message", "new feature", "--allow-empty")
+def test_creating_merge_request(repo_path, tsrc_cli, gitlab_mock, push_args):
+    tsrc.git.run_git(repo_path, "checkout", "-b", "new-feature")
+    tsrc.git.run_git(repo_path, "commit", "--message", "new feature", "--allow-empty")
 
     gitlab_mock.find_opened_merge_request.return_value = list()
     gitlab_mock.create_merge_request.return_value = MR_STUB
@@ -90,7 +67,7 @@ def test_creating_merge_request(foo_path, tsrc_cli, gitlab_mock, push_args):
     push_args.target_branch = "next"
     push_args.mr_title = "Best feature ever"
 
-    execute_push(foo_path, push_args, gitlab_mock)
+    execute_push(repo_path, push_args, gitlab_mock)
 
     gitlab_mock.assert_mr_created(
         "42", "new-feature",
@@ -106,15 +83,15 @@ def test_creating_merge_request(foo_path, tsrc_cli, gitlab_mock, push_args):
     )
 
 
-def test_existing_merge_request(foo_path, tsrc_cli, gitlab_mock, push_args):
-    tsrc.git.run_git(foo_path, "checkout", "-b", "new-feature")
-    tsrc.git.run_git(foo_path, "commit", "--message", "new feature", "--allow-empty")
+def test_existing_merge_request(repo_path, tsrc_cli, gitlab_mock, push_args):
+    tsrc.git.run_git(repo_path, "checkout", "-b", "new-feature")
+    tsrc.git.run_git(repo_path, "commit", "--message", "new feature", "--allow-empty")
 
     gitlab_mock.find_opened_merge_request.return_value = MR_STUB
 
     push_args.target_branch = "next"
     push_args.mr_title = "Best feature ever"
-    execute_push(foo_path, push_args, gitlab_mock)
+    execute_push(repo_path, push_args, gitlab_mock)
 
     gitlab_mock.assert_mr_not_created()
     gitlab_mock.assert_mr_updated(
@@ -125,19 +102,19 @@ def test_existing_merge_request(foo_path, tsrc_cli, gitlab_mock, push_args):
     )
 
 
-def test_accept_merge_request(foo_path, tsrc_cli, gitlab_mock, push_args):
-    tsrc.git.run_git(foo_path, "checkout", "-b", "new-feature")
-    tsrc.git.run_git(foo_path, "commit", "--message", "new feature", "--allow-empty")
+def test_accept_merge_request(repo_path, tsrc_cli, gitlab_mock, push_args):
+    tsrc.git.run_git(repo_path, "checkout", "-b", "new-feature")
+    tsrc.git.run_git(repo_path, "commit", "--message", "new feature", "--allow-empty")
 
     gitlab_mock.find_opened_merge_request.return_value = MR_STUB
 
     push_args.accept = True
-    execute_push(foo_path, push_args, gitlab_mock)
+    execute_push(repo_path, push_args, gitlab_mock)
 
     gitlab_mock.assert_mr_accepted(MR_STUB)
 
 
-def test_unwipify_existing_merge_request(foo_path, tsrc_cli, gitlab_mock, push_args):
+def test_unwipify_existing_merge_request(repo_path, tsrc_cli, gitlab_mock, push_args):
     existing_mr = {
         "title": "WIP: nice title",
         "web_url": "http://example.com/42",
@@ -146,7 +123,7 @@ def test_unwipify_existing_merge_request(foo_path, tsrc_cli, gitlab_mock, push_a
     gitlab_mock.find_opened_merge_request.return_value = existing_mr
 
     push_args.ready = True
-    execute_push(foo_path, push_args, gitlab_mock)
+    execute_push(repo_path, push_args, gitlab_mock)
 
     gitlab_mock.assert_mr_not_created()
     gitlab_mock.assert_mr_updated(
