+++ tsrc/test/helpers/push.py
@@ -1,0 +1,29 @@
+import types
+
+import pytest
+
+
+@pytest.fixture
+def push_args():
+    args = types.SimpleNamespace()
+    args.accept = False
+    args.assignee = None
+    args.force = False
+    args.merge = False
+    args.mr_title = None
+    args.ready = False
+    args.target_branch = "master"
+    args.title = None
+    args.wip = False
+    return args
+
+
+@pytest.fixture
+def repo_path(monkeypatch, git_server, tsrc_cli, workspace_path):
+    """ Path to a freshly cloned repository """
+    git_server.add_repo("foo/bar")
+    manifest_url = git_server.manifest_url
+    tsrc_cli.run("init", manifest_url)
+    foo_path = workspace_path.joinpath("foo/bar")
+    monkeypatch.chdir(foo_path)
+    return foo_path
